refactor into installable package
