+++ .gitignore
@@ -1,1 +1,3 @@
 *.pyc
+KevinScript.egg-info
+.idea
