+++ ks/__init__.py
@@ -1,32 +1,21 @@
-from __future__ import print_function
-# ugly stuff to import modules one directory up
+import StringIO
 import os
 import sys
-try:
-    from StringIO import StringIO
-except ImportError:
-    from io import StringIO
+
+from ks.eval_ast import NodeConstructor, evaluate
+from ks.parser import ast, parserExceptions
 
 cur_dir = os.path.dirname(os.path.abspath(__file__))
 top_dir = os.path.dirname(cur_dir)
-parser_dir = os.path.join(top_dir, "lib", "parser")
-sys.path.insert(0, parser_dir)
-import ast
-import parserExceptions
-from eval_ast import evaluate, NodeConstructor
-
-try:
-    input = raw_input
-except:
-    pass
 
 reducible_nodes = ["StatementList", "ExpressionList", "IdentifierList", "KeyValueList", "FunctionDeclarationStatementList"]
 
 base_compile = ast.get_compiler(
-    os.path.join(cur_dir, "tokens.txt"), 
-    os.path.join(cur_dir, "language.txt"), 
+    os.path.join(cur_dir, "tokens.txt"),
+    os.path.join(cur_dir, "language.txt"),
     reducible_nodes
 )
+
 
 def compile(program_text, strict=False):
     try:
@@ -40,18 +29,20 @@ def compile(program_text, strict=False):
             raise
     return tree
 
+
 def execute(program_text, strict=False, mode="exec"):
     assert mode in ["exec", "single"], "did not recognize execution mode '{}'".format(mode)
     tree = compile(program_text, strict)
     if mode == "single":
         #primarily used by the REPL. if the final statement in the program is an expression that doesn't evaluate to None, print its result.
-        #print("entering pdb...")
+        #print "entering pdb..."
         #import pdb; pdb.set_trace()
         assert tree.klass == "StatementList"
         last_statement = tree.children[-1].children[0]
         if last_statement.klass == "ExpressionStatement":
             last_statement.children[0] = NodeConstructor.make_identifier_call("print_single", [last_statement.children[0]])
     evaluate(tree)
+
 
 def check_output(*args, **kargs):
     """
@@ -67,7 +58,7 @@ def check_output(*args, **kargs):
                 channel.write(data)
 
     verbose = kargs.pop("verbose", True)
-    string_io = StringIO()
+    string_io = StringIO.StringIO()
     old_stdout = sys.stdout
 
     if verbose:
@@ -81,16 +72,16 @@ def check_output(*args, **kargs):
         sys.stdout = old_stdout
 
     return string_io.getvalue().rstrip()
-    
+
 
 def repl():
     def isEofException(ex):
         return isinstance(ex, parserExceptions.NoActionFoundError) and ex.token.klass.name == "$"
     data = ""
-    print(">>>", end = " ")
+    print ">>>",
     while True:
         try:
-            line = input("")
+            line = raw_input("")
         except (EOFError, KeyboardInterrupt):
             return
         data += line
@@ -99,19 +90,23 @@ def repl():
             execute(data, strict=bool(line), mode="single")
         except Exception as ex:
             if isEofException(ex):
-                print("...", end = " ")
+                print "...",
                 continue
             else:
-                print(ex)
+                print ex
         data = ""
-        print(">>>", end = " ")
+        print ">>>",
+
 
 with open(os.path.join(cur_dir, "native_builtin_initialization.k")) as file:
     execute(file.read())
 
-if __name__ == "__main__":
-    import sys
-    if len(sys.argv) < 2:
+
+def main(*argv):
+    if not argv:
+        argv = sys.argv
+
+    if len(argv) < 2:
         repl()
         sys.exit(0)
