+++ setup.cfg
@@ -1,0 +1,3 @@
+[egg_info]
+tag_date=1
+tag_build=.dev
